[jackson] Report mismatches in order
By switching from a Map or LinkedHashMap we can maintain insertion order.
This is helpful when reporting a mismatch in two ways.
First it is consistent with what the user has entered, and second we can properly
print out the ellipsis only when there is material that matches in the middle of mismatches.